Completed member registration, going to add a list for those who applied.
