+++ excelr8/excelr8/urls.py
@@ -24,7 +24,8 @@ urlpatterns = [
     path('4dm1n_d4shb04rd_142004/', admin.site.urls),
     path('', include('home.urls')),
     path('members/', include('team_management.urls')),
-    path('4dm1n_d4shb04rd_3987234098274091823712931/', include('admin_dashboard.urls'))
+    path('4dm1n_d4shb04rd_3987234098274091823712931/', include('admin_dashboard.urls')),
+    # path('register/', include('team_management.urls.register'), name='register'),
 ]
 
 if settings.DEBUG:
