+++ excelr8/home/templates/about.html
@@ -45,7 +45,7 @@
     </div>
     <div class="about-section">
         <h2 style="color: cyan;">Join Us</h2>
-        <p>Are you passionate about cybersecurity, programming, or tech research? Be a part of our study group and let's make a difference together! <a href="{% url 'register' %}">Register now!</a></p>
+        <p>Love cybersecurity, programming, or tech research? Join us and be part of something awesome! <a href="{% url 'team_management:register_member' %}">Apply now!</a></p>
     </div>
 </div>
