+++ excelr8/home/templates/base.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="icon" href="{% static 'home/favicon.ico' %}" type="image/x-icon" sizes="64x64">  
+    <link rel="icon" href="{% static 'home/favicon.ico' %}" type="image/x-icon" sizes="128x128">  
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Courier+Prime&display=swap" rel="stylesheet">
     <style>
