+++ excelr8/home/urls.py
@@ -6,5 +6,4 @@ urlpatterns = [
     path('about/', views.about, name='about'),
     path('publications/', views.publications, name='publications'),
     path('projects/', views.projects, name='projects'),
-    path('register/', views.register, name='register'),
-]
+    ]
