+++ excelr8/team_management/admin.py
@@ -1,3 +1,4 @@
 from django.contrib import admin
+from .models import Member  # Replace with the actual model name from models.py
 
-# Register your models here.
+admin.site.register(Member)  # Register the model from models.py here
