+++ excelr8/team_management/urls.py
@@ -8,7 +8,8 @@ app_name = 'team_management'
 urlpatterns = [
     path('', views.team_list, name='team_list'),
     path('<int:member_id>/', views.member_detail, name='member_detail'),
-    path('register/', views.register, name='register'),
+    path('register_member/', views.register, name='register_member'),
+    
 ]
 
 if settings.DEBUG:
