+++ excelr8/team_management/views.py
@@ -1,21 +1,18 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
 from admin_dashboard.models import TeamMember 
-
+from .models import Member 
+ 
 def team_list(request):
-    members = TeamMember.objects.all()  # Fetch members from the database
-    return render(request, 'team_list.html', {'team_members': members})  # Pass members to the template
+    members = TeamMember.objects.all()
+    return render(request, 'team_list.html', {'team_members': members}) 
 
 def member_detail(request, member_id):
     member = get_object_or_404(TeamMember, pk=member_id)
     return render(request, 'member_detail.html', {'member': member})
 
 def register(request):
-    from .models import Member
-    from django.shortcuts import redirect
-
     if request.method == 'POST':
-        # Extract form data
         first_name = request.POST.get('first_name')
         last_name = request.POST.get('last_name')
         email = request.POST.get('email')
@@ -25,7 +22,6 @@ def register(request):
         interests = request.POST.getlist('interests')
         other_interest = request.POST.get('other_interest', '')
 
-        # Create a new member object and save it to the database
         new_member = Member(
             first_name=first_name,
             last_name=last_name,
@@ -37,10 +33,7 @@ def register(request):
             other_interest=other_interest
         )
         new_member.save()
+        messages.success(request, 'Registration successful!') 
+        return render(request, 'member_registration.html', {'success': True}) 
 
-        # Redirect to a success page
-        return redirect('registration_success')  # Assuming there's a URL named 'registration_success'
-
-    else:
-        # If a GET (or any other method) we'll create a blank form
-        return render(request, 'member_registration.html')
+    return render(request, 'member_registration.html')
